Reorganized code to highlight standard and compat classes
