+++ componentapiexample/src/main/java/net/gini/android/vision/component/analysis/compat/AnalysisExampleAppCompatActivity.java
@@ -1,22 +1,22 @@
-package net.gini.android.vision.component.analysis;
+package net.gini.android.vision.component.analysis.compat;
 
-import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
+import android.support.v7.app.AppCompatActivity;
 
 import net.gini.android.vision.Document;
 import net.gini.android.vision.GiniVisionError;
 import net.gini.android.vision.analysis.AnalysisFragmentListener;
 import net.gini.android.vision.component.R;
 
-public class AnalysisExampleActivity extends Activity implements
+public class AnalysisExampleAppCompatActivity extends AppCompatActivity implements
         AnalysisFragmentListener {
 
     public static final String EXTRA_IN_DOCUMENT = "EXTRA_IN_DOCUMENT";
     public static final String EXTRA_IN_ERROR_MESSAGE = "EXTRA_IN_ERROR_MESSAGE";
-    private AnalysisScreenHandler mAnalysisScreenHandler;
+    private AnalysisScreenHandlerAppCompat mAnalysisScreenHandler;
 
     @Override
     public void onAnalyzeDocument(@NonNull final Document document) {
@@ -31,14 +31,14 @@ public class AnalysisExampleActivity extends Activity implements
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_analysis);
-        mAnalysisScreenHandler = new AnalysisScreenHandler(this);
+        setContentView(R.layout.activity_analysis_compat);
+        mAnalysisScreenHandler = new AnalysisScreenHandlerAppCompat(this);
         mAnalysisScreenHandler.onCreate(savedInstanceState);
     }
 
     public static Intent newInstance(final Document document,
             final String errorMessage, final Context context) {
-        final Intent intent = new Intent(context, AnalysisExampleActivity.class);
+        final Intent intent = new Intent(context, AnalysisExampleAppCompatActivity.class);
         intent.putExtra(EXTRA_IN_DOCUMENT, document);
         intent.putExtra(EXTRA_IN_ERROR_MESSAGE, errorMessage);
         return intent;
