+++ componentapiexample/src/main/java/net/gini/android/vision/component/review/standard/ReviewExampleActivity.java
@@ -1,35 +1,35 @@
-package net.gini.android.vision.component.review;
+package net.gini.android.vision.component.review.standard;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
 
 import net.gini.android.vision.Document;
 import net.gini.android.vision.GiniVisionError;
 import net.gini.android.vision.component.R;
 import net.gini.android.vision.review.ReviewFragmentListener;
 
-public class ReviewExampleAppCompatActivity extends AppCompatActivity implements
+public class ReviewExampleActivity extends Activity implements
         ReviewFragmentListener {
 
     public static final String EXTRA_IN_DOCUMENT = "EXTRA_IN_DOCUMENT";
-    private ReviewScreenHandlerAppCompat mReviewScreenHandler;
+    private ReviewScreenHandler mReviewScreenHandler;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_review);
+        mReviewScreenHandler = new ReviewScreenHandler(this);
+        mReviewScreenHandler.onCreate(savedInstanceState);
+    }
 
     @Override
     protected void onActivityResult(final int requestCode, final int resultCode,
             final Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
         mReviewScreenHandler.onActivityResult(requestCode, resultCode, data);
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_review_compat);
-        mReviewScreenHandler = new ReviewScreenHandlerAppCompat(this);
-        mReviewScreenHandler.onCreate(savedInstanceState);
     }
 
     @Override
@@ -59,8 +59,8 @@ public class ReviewExampleAppCompatActivity extends AppCompatActivity implements
     }
 
     public static Intent newInstance(final Document document, final Context context) {
-        final Intent intent = new Intent(context, ReviewExampleAppCompatActivity.class);
-        intent.putExtra(ReviewExampleAppCompatActivity.EXTRA_IN_DOCUMENT, document);
+        final Intent intent = new Intent(context, ReviewExampleActivity.class);
+        intent.putExtra(ReviewExampleActivity.EXTRA_IN_DOCUMENT, document);
         return intent;
     }
 }
